requirements.txt
flask
python-dotenv

langchain
langchain_community
langchain-google-genai


sentence-transformers

pinecone-client

pypdf

google-generativeai
